src/disease_model/requirements.txt
pandas~=0.25.3
gitpython~=3.1.2
numpy~=1.17.5
scipy~=1.3.3